requirements.txt
pendulum>=3.0.0
requests>=2.31.0
apache-airflow-providers-http
apache-airflow-providers-google
apache-airflow-providers-postgres
pandas>=2.0.3